Second With Documentation and Printing
The goal of the calculator was to determine the day of the week the
user was born. This last part of code was the missing piece. The
autcome is notified via alert, displayed on the webpage and printed on
paper with all the other document.write elements on the program written
before.